+++ src/frontend/src/features/rooms/livekit/components/effects/Effects.tsx
@@ -1,15 +1,14 @@
 import { useLocalParticipant } from '@livekit/components-react'
-import { LocalVideoTrack } from 'livekit-client'
+import { LocalVideoTrack, Track } from 'livekit-client'
 import { css } from '@/styled-system/css'
 import { EffectsConfiguration } from './EffectsConfiguration'
 import { useCanPublishTrack } from '@/features/rooms/livekit/hooks/useCanPublishTrack'
-import { TrackSource } from '@livekit/protocol'
 
 export const Effects = () => {
   const { cameraTrack } = useLocalParticipant()
   const localCameraTrack = cameraTrack?.track as LocalVideoTrack
 
-  const canPublishCamera = useCanPublishTrack(TrackSource.CAMERA)
+  const canPublishCamera = useCanPublishTrack(Track.Source.Camera)
 
   return (
     <div